output of clang90
Interesting that clang has looked for a formal location to output the object file. Some of the headers were not present, so I skipped them. Unknown arguments and missing headers could be construed. This really is the first run of clang after several nights of searching files, and then a few runs with GCC. Clearly I was getting nowhere with GCC. Pcpu.h will not run with GCC.